+++ extensions/mapstore/src/main/java/com/hazelcast/mapstore/JdbcParameters.java
@@ -22,18 +22,40 @@ import com.hazelcast.sql.SqlColumnMetadata;
 
 import java.util.List;
 
-final class ToJDBCParametersConverter {
+class JdbcParameters {
+    private int idPos;
 
-    private ToJDBCParametersConverter() {
+    private Object[] params;
+
+    void setIdPos(int idPos) {
+        this.idPos = idPos;
+    }
+
+    Object[] getParams() {
+        return params;
+    }
+
+    void setParams(Object[] params) {
+        this.params = params;
+    }
+
+    void shiftIdParameterToEnd() {
+        Object id = params[idPos];
+        for (int i = idPos; i < params.length - 1; i++) {
+            params[i] = params[i + 1];
+        }
+        params[params.length - 1] = id;
     }
 
     // Convert key and GenericRecord to JDBC parameter values
-    static <K> JDBCParameters convert(K key,
-                                      GenericRecord genericRecord,
-                                      List<SqlColumnMetadata> columnMetadataList,
-                                      String idColumn) {
+    static <K> JdbcParameters convert(
+            K key,
+            GenericRecord genericRecord,
+            List<SqlColumnMetadata> columnMetadataList,
+            String idColumn
+    ) {
 
-        JDBCParameters jdbcParameters = new JDBCParameters();
+        JdbcParameters jdbcParameters = new JdbcParameters();
 
         int idPos = -1;
         Object[] params = new Object[columnMetadataList.size()];
@@ -116,4 +138,5 @@ final class ToJDBCParametersConverter {
 
         return jdbcParameters;
     }
+
 }
