--- extensions/mapstore/src/main/java/com/hazelcast/mapstore/validators/package-info.java
@@ -1,20 +1,0 @@
-/*
- * Copyright 2023 Hazelcast Inc.
- *
- * Licensed under the Hazelcast Community License (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://hazelcast.com/hazelcast-community-license
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Contains classes related to validation
- */
-package com.hazelcast.mapstore.validators;
